fix(feedback): restore unread badge by switching context polling to direct API
